Cleaned up a lot of code

+++ src/main/java/com/mjr/extraplanets/blocks/BlockNuclearBomb.java
@@ -29,9 +29,9 @@ public class BlockNuclearBomb extends Block{
 
     public BlockNuclearBomb()
     {
-        super(Material.tnt);
-        this.setCreativeTab(ExtraPlanets.BlocksTab);
-        this.setBlockName("nuclearBomb");
+	super(Material.tnt);
+	this.setCreativeTab(ExtraPlanets.BlocksTab);
+	this.setBlockName("nuclearBomb");
     }
 
     /**
@@ -41,7 +41,7 @@ public class BlockNuclearBomb extends Block{
     @SideOnly(Side.CLIENT)
     public IIcon getIcon(int p_149691_1_, int p_149691_2_)
     {
-        return p_149691_1_ == 0 ? this.field_150115_b : (p_149691_1_ == 1 ? this.field_150116_a : this.blockIcon);
+	return p_149691_1_ == 0 ? this.field_150115_b : (p_149691_1_ == 1 ? this.field_150116_a : this.blockIcon);
     }
 
     /**
@@ -50,13 +50,13 @@ public class BlockNuclearBomb extends Block{
     @Override
     public void onBlockAdded(World p_149726_1_, int p_149726_2_, int p_149726_3_, int p_149726_4_)
     {
-        super.onBlockAdded(p_149726_1_, p_149726_2_, p_149726_3_, p_149726_4_);
+	super.onBlockAdded(p_149726_1_, p_149726_2_, p_149726_3_, p_149726_4_);
 
-        if (p_149726_1_.isBlockIndirectlyGettingPowered(p_149726_2_, p_149726_3_, p_149726_4_))
-        {
-            this.onBlockDestroyedByPlayer(p_149726_1_, p_149726_2_, p_149726_3_, p_149726_4_, 1);
-            p_149726_1_.setBlockToAir(p_149726_2_, p_149726_3_, p_149726_4_);
-        }
+	if (p_149726_1_.isBlockIndirectlyGettingPowered(p_149726_2_, p_149726_3_, p_149726_4_))
+	{
+	    this.onBlockDestroyedByPlayer(p_149726_1_, p_149726_2_, p_149726_3_, p_149726_4_, 1);
+	    p_149726_1_.setBlockToAir(p_149726_2_, p_149726_3_, p_149726_4_);
+	}
     }
 
     /**
@@ -66,11 +66,11 @@ public class BlockNuclearBomb extends Block{
     @Override
     public void onNeighborBlockChange(World p_149695_1_, int p_149695_2_, int p_149695_3_, int p_149695_4_, Block p_149695_5_)
     {
-        if (p_149695_1_.isBlockIndirectlyGettingPowered(p_149695_2_, p_149695_3_, p_149695_4_))
-        {
-            this.onBlockDestroyedByPlayer(p_149695_1_, p_149695_2_, p_149695_3_, p_149695_4_, 1);
-            p_149695_1_.setBlockToAir(p_149695_2_, p_149695_3_, p_149695_4_);
-        }
+	if (p_149695_1_.isBlockIndirectlyGettingPowered(p_149695_2_, p_149695_3_, p_149695_4_))
+	{
+	    this.onBlockDestroyedByPlayer(p_149695_1_, p_149695_2_, p_149695_3_, p_149695_4_, 1);
+	    p_149695_1_.setBlockToAir(p_149695_2_, p_149695_3_, p_149695_4_);
+	}
     }
 
     /**
@@ -79,7 +79,7 @@ public class BlockNuclearBomb extends Block{
     @Override
     public int quantityDropped(Random p_149745_1_)
     {
-        return 1;
+	return 1;
     }
 
     /**
@@ -88,12 +88,12 @@ public class BlockNuclearBomb extends Block{
     @Override
     public void onBlockDestroyedByExplosion(World p_149723_1_, int p_149723_2_, int p_149723_3_, int p_149723_4_, Explosion p_149723_5_)
     {
-        if (!p_149723_1_.isRemote)
-        {
-            EntityNuclearBombPrimed entitynuclearBombprimed = new EntityNuclearBombPrimed(p_149723_1_, (double)((float)p_149723_2_ + 0.5F), (double)((float)p_149723_3_ + 0.5F), (double)((float)p_149723_4_ + 0.5F), p_149723_5_.getExplosivePlacedBy());
-            entitynuclearBombprimed.fuse = p_149723_1_.rand.nextInt(entitynuclearBombprimed.fuse / 4) + entitynuclearBombprimed.fuse / 8;
-            p_149723_1_.spawnEntityInWorld(entitynuclearBombprimed);
-        }
+	if (!p_149723_1_.isRemote)
+	{
+	    EntityNuclearBombPrimed entitynuclearBombprimed = new EntityNuclearBombPrimed(p_149723_1_, p_149723_2_ + 0.5F, p_149723_3_ + 0.5F, p_149723_4_ + 0.5F, p_149723_5_.getExplosivePlacedBy());
+	    entitynuclearBombprimed.fuse = p_149723_1_.rand.nextInt(entitynuclearBombprimed.fuse / 4) + entitynuclearBombprimed.fuse / 8;
+	    p_149723_1_.spawnEntityInWorld(entitynuclearBombprimed);
+	}
     }
 
     /**
@@ -102,20 +102,20 @@ public class BlockNuclearBomb extends Block{
     @Override
     public void onBlockDestroyedByPlayer(World p_149664_1_, int p_149664_2_, int p_149664_3_, int p_149664_4_, int p_149664_5_)
     {
-        this.primeTnt(p_149664_1_, p_149664_2_, p_149664_3_, p_149664_4_, p_149664_5_, (EntityLivingBase)null);
+	this.primeTnt(p_149664_1_, p_149664_2_, p_149664_3_, p_149664_4_, p_149664_5_, (EntityLivingBase)null);
     }
-    
+
     public void primeTnt(World p_150114_1_, int p_150114_2_, int p_150114_3_, int p_150114_4_, int p_150114_5_, EntityLivingBase p_150114_6_)
     {
-        if (!p_150114_1_.isRemote)
-        {
-            if ((p_150114_5_ & 1) == 1)
-            {
-        	EntityNuclearBombPrimed entitynuclearBombprimed = new EntityNuclearBombPrimed(p_150114_1_, (double)((float)p_150114_2_ + 0.5F), (double)((float)p_150114_3_ + 0.5F), (double)((float)p_150114_4_ + 0.5F), p_150114_6_);
-                p_150114_1_.spawnEntityInWorld(entitynuclearBombprimed);
-                p_150114_1_.playSoundAtEntity(entitynuclearBombprimed, "game.tnt.primed", 1.0F, 1.0F);
-            }
-        }
+	if (!p_150114_1_.isRemote)
+	{
+	    if ((p_150114_5_ & 1) == 1)
+	    {
+		EntityNuclearBombPrimed entitynuclearBombprimed = new EntityNuclearBombPrimed(p_150114_1_, p_150114_2_ + 0.5F, p_150114_3_ + 0.5F, p_150114_4_ + 0.5F, p_150114_6_);
+		p_150114_1_.spawnEntityInWorld(entitynuclearBombprimed);
+		p_150114_1_.playSoundAtEntity(entitynuclearBombprimed, "game.tnt.primed", 1.0F, 1.0F);
+	    }
+	}
     }
 
     /**
@@ -124,17 +124,17 @@ public class BlockNuclearBomb extends Block{
     @Override
     public boolean onBlockActivated(World p_149727_1_, int p_149727_2_, int p_149727_3_, int p_149727_4_, EntityPlayer p_149727_5_, int p_149727_6_, float p_149727_7_, float p_149727_8_, float p_149727_9_)
     {
-        if (p_149727_5_.getCurrentEquippedItem() != null && p_149727_5_.getCurrentEquippedItem().getItem() == Items.flint_and_steel)
-        {
-            this.primeTnt(p_149727_1_, p_149727_2_, p_149727_3_, p_149727_4_, 1, p_149727_5_);
-            p_149727_1_.setBlockToAir(p_149727_2_, p_149727_3_, p_149727_4_);
-            p_149727_5_.getCurrentEquippedItem().damageItem(1, p_149727_5_);
-            return true;
-        }
-        else
-        {
-            return super.onBlockActivated(p_149727_1_, p_149727_2_, p_149727_3_, p_149727_4_, p_149727_5_, p_149727_6_, p_149727_7_, p_149727_8_, p_149727_9_);
-        }
+	if (p_149727_5_.getCurrentEquippedItem() != null && p_149727_5_.getCurrentEquippedItem().getItem() == Items.flint_and_steel)
+	{
+	    this.primeTnt(p_149727_1_, p_149727_2_, p_149727_3_, p_149727_4_, 1, p_149727_5_);
+	    p_149727_1_.setBlockToAir(p_149727_2_, p_149727_3_, p_149727_4_);
+	    p_149727_5_.getCurrentEquippedItem().damageItem(1, p_149727_5_);
+	    return true;
+	}
+	else
+	{
+	    return super.onBlockActivated(p_149727_1_, p_149727_2_, p_149727_3_, p_149727_4_, p_149727_5_, p_149727_6_, p_149727_7_, p_149727_8_, p_149727_9_);
+	}
     }
 
     /**
@@ -143,16 +143,16 @@ public class BlockNuclearBomb extends Block{
     @Override
     public void onEntityCollidedWithBlock(World p_149670_1_, int p_149670_2_, int p_149670_3_, int p_149670_4_, Entity p_149670_5_)
     {
-        if (p_149670_5_ instanceof EntityArrow && !p_149670_1_.isRemote)
-        {
-            EntityArrow entityarrow = (EntityArrow)p_149670_5_;
+	if (p_149670_5_ instanceof EntityArrow && !p_149670_1_.isRemote)
+	{
+	    EntityArrow entityarrow = (EntityArrow)p_149670_5_;
 
-            if (entityarrow.isBurning())
-            {
-                this.primeTnt(p_149670_1_, p_149670_2_, p_149670_3_, p_149670_4_, 1, entityarrow.shootingEntity instanceof EntityLivingBase ? (EntityLivingBase)entityarrow.shootingEntity : null);
-                p_149670_1_.setBlockToAir(p_149670_2_, p_149670_3_, p_149670_4_);
-            }
-        }
+	    if (entityarrow.isBurning())
+	    {
+		this.primeTnt(p_149670_1_, p_149670_2_, p_149670_3_, p_149670_4_, 1, entityarrow.shootingEntity instanceof EntityLivingBase ? (EntityLivingBase)entityarrow.shootingEntity : null);
+		p_149670_1_.setBlockToAir(p_149670_2_, p_149670_3_, p_149670_4_);
+	    }
+	}
     }
 
     /**
@@ -161,14 +161,14 @@ public class BlockNuclearBomb extends Block{
     @Override
     public boolean canDropFromExplosion(Explosion p_149659_1_)
     {
-        return false;
+	return false;
     }
     @Override
     @SideOnly(Side.CLIENT)
     public void registerBlockIcons(IIconRegister p_149651_1_)
     {
-        this.blockIcon = p_149651_1_.registerIcon(Constants.TEXTURE_PREFIX + "nuclearBomb" + "_side");
-        this.field_150116_a = p_149651_1_.registerIcon(Constants.TEXTURE_PREFIX + "nuclearBomb" + "_top");
-        this.field_150115_b = p_149651_1_.registerIcon(Constants.TEXTURE_PREFIX + "nuclearBomb" + "_bottom");
+	this.blockIcon = p_149651_1_.registerIcon(Constants.TEXTURE_PREFIX + "nuclearBomb" + "_side");
+	this.field_150116_a = p_149651_1_.registerIcon(Constants.TEXTURE_PREFIX + "nuclearBomb" + "_top");
+	this.field_150115_b = p_149651_1_.registerIcon(Constants.TEXTURE_PREFIX + "nuclearBomb" + "_bottom");
     }
 }
