+++ src/main/java/com/mjr/extraplanets/entities/EntityNuclearBombPrimed.java
@@ -24,9 +24,9 @@ public class EntityNuclearBombPrimed extends Entity {
 	System.out.println("here1");
 	this.setPosition(p_i1730_2_, p_i1730_4_, p_i1730_6_);
 	float f = (float) (Math.random() * Math.PI * 2.0D);
-	this.motionX = (double) (-((float) Math.sin((double) f)) * 0.02F);
+	this.motionX = -((float) Math.sin(f)) * 0.02F;
 	this.motionY = 0.20000000298023224D;
-	this.motionZ = (double) (-((float) Math.cos((double) f)) * 0.02F);
+	this.motionZ = -((float) Math.cos(f)) * 0.02F;
 	this.fuse = 80;
 	this.prevPosX = p_i1730_2_;
 	this.prevPosY = p_i1730_4_;
